+++ expediagroup-sdk-graphql/src/main/kotlin/com/expediagroup/sdk/graphql/GraphQLExecutor.kt
@@ -21,8 +21,8 @@ import com.apollographql.apollo.api.Operation
 import com.expediagroup.sdk.core.exception.service.ExpediaGroupServiceException
 import com.expediagroup.sdk.core.transport.AbstractRequestExecutor
 import com.expediagroup.sdk.graphql.exception.NoDataException
-import com.expediagroup.sdk.graphql.model.Error
-import com.expediagroup.sdk.graphql.model.RawResponse
+import com.expediagroup.sdk.graphql.model.GraphQLError
+import com.expediagroup.sdk.graphql.model.RawGraphQLResponse
 
 /**
  * A streamlined implementation of [GraphQLExecutor] that handles GraphQL operations with
@@ -40,14 +40,14 @@ class GraphQLExecutor(
 ) : AbstractGraphQLExecutor(requestExecutor) {
 
     /**
-     * Executes a GraphQL operation and returns a [RawResponse] containing the complete data and any errors.
+     * Executes a GraphQL operation and returns a [RawGraphQLResponse] containing the complete data and any errors.
      *
      * @param operation The GraphQL operation to be executed.
-     * @return A [RawResponse] with the full data structure and any errors from the server.
+     * @return A [RawGraphQLResponse] with the full data structure and any errors from the server.
      * @throws [ExpediaGroupServiceException] If an exception occurs during operation execution.
      * @throws [NoDataException] If the operation completes without data but includes errors.
      */
-    override fun <T : Operation.Data> execute(operation: Operation<T>): RawResponse<T> = operation
+    override fun <T : Operation.Data> execute(operation: Operation<T>): RawGraphQLResponse<T> = operation
         .toSDKRequest(serverUrl).let {
             requestExecutor.execute(it)
         }.toApolloResponse(operation).let {
@@ -59,18 +59,18 @@ class GraphQLExecutor(
      *
      * @param response The ApolloResponse containing the data and errors from the GraphQL operation.
      */
-    private fun <T : Operation.Data> processApolloResponse(response: ApolloResponse<T>): RawResponse<T> {
+    private fun <T : Operation.Data> processApolloResponse(response: ApolloResponse<T>): RawGraphQLResponse<T> {
         return when {
             response.exception != null -> throw ExpediaGroupServiceException(cause = response.exception)
 
             response.data == null && response.hasErrors() -> throw NoDataException(
                 message = "No data received from the server",
-                errors = response.errors!!.map { Error.fromApolloError(it) }
+                errors = response.errors!!.map { GraphQLError.fromApolloError(it) }
             )
 
-            else -> RawResponse(
+            else -> RawGraphQLResponse(
                 data = response.data!!,
-                errors = response.errors?.map { Error.fromApolloError(it) }
+                errors = response.errors?.map { GraphQLError.fromApolloError(it) }
             )
         }
     }
