+++ expediagroup-sdk-graphql/src/main/kotlin/com/expediagroup/sdk/graphql/paging/PageInfo.kt
