+++ settings.gradle
@@ -9,5 +9,5 @@ include 'examples'
 include 'apollo-compiler-plugin'
 include 'expediagroup-sdk-core'
 include 'expediagroup-sdk-okhttp-transport'
-include 'expediagroup-sdk-graphql-support'
+include 'expediagroup-sdk-graphql'
